ciaos/requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
google-api-python-client==1.6.5
gspread==0.6.2
httplib2==0.10.3
idna==2.6
oauth2client==4.1.2
Pillow==5.0.0
pyaes==1.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyOpenSSL==17.5.0
python-gflags==3.1.2
requests==2.18.4
rsa==3.4.2
six==1.11.0
Telethon==0.16.2.2
uritemplate==3.0.0
urllib3==1.22
gunicorn==19.7.1
